refactor(salarié): Réorganise les composants de fiche de paie
