+++ site/source/components/FicheDePaie/PaySlip.css
